A lot of minor modifications, and make clear that Vorbis is mandatory
Heya,

Here are a couple of fixes to the two sound specs. Nothing too
interesting, except that I declared Vorbis support mandatory. Why?
Because theme designers and applications for the desktop should be
able to rely on that Vorbis support is available. I want to encourage
people to ship their sound themes in Vorbis.

Nobody forces Nokia to follow this spec 100%. I mean, fd.o is not
really a standardization institution, so our standards are not worth
too much anyway.

So, if Nokia really hates Vorbis that much than they are welcome to
implement this spec only partially.

I think I am mostly happy now for a release 0.1. I'd like to see
email-message-new and instant-message-new changed to message-new-email
and message-new-instant however, since it makes sense to share a
common sound for them. However, I've been too lazy to do all this
swapping foo. Will do that in another patch, maybe.

Lennart

Signed-off-by: Marc-Andre Lureau <marcandre.lureau@gmail.com>