mime-apps: firm up the spec in some places
 - the MimeType= line in .desktop files in ~ takes precedence over
   mimeapps.list in lower-priority directories

 - firm up the interaction between "Added" and "Default" (which is
   almost nothing now, except that we fall back to the list of all
   supported applications in case we don't find a default)

 - specify that the order of the Added list should be preserved

I've also changed the order of the associations and defaults section in
the document to improve ease of reading.  Defaults depends on
associations (for the case where a default is not found) but
associations make no mention of defaults.  It is also helpful for
allowing mention of the fact that higher-level desktop files do NOT
cause masking in the case of defaults (in contrast to the mention in the
previous section, where the masking behaviour is specified).