mime-apps: change rules for default apps
Reverse the earlier language that selecting default apps has nothing to
do with associations.  We now require that an app is associated with a
particular type before it can be considered as the default for that
type.

This resolves a couple of tricky cases in terms of expected behaviour
when removing the association between an app and a particular type.

 - there is no need to explicitly select a "new default" at this point
   which may become uninstalled in the future (which would result in
   fallback to the application that we removed the association for)

 - it is now possible to have 'no default app' in case the only app for
   a given mime type had its association manually removed

This does add an additional requirement on implementations, however:

 - implementations should ensure that a type association exists when
   between an app and a type when setting that app as the default

https://bugs.freedesktop.org/show_bug.cgi?id=86069