desktop-entry-spec: document DesktopNames
Document the DesktopNames key as the mechanism by which
the XDG_CURRENT_DESKTOP environment variable is expected to be set.

We don't document this as a key in the normal way, because it is not a
valid key for desktop files.  We only include a mention of it here
because the desktop entry spec is more or less the "home" of what
XDG_CURRENT_DESKTOP means, and the mechanism by which it is normally set
should be part of that definition.

https://bugs.freedesktop.org/show_bug.cgi?id=85938