desktop-entry-spec: allow '-' in application ids
Up to now, the spec has not clearly specified what will happen if a '-'
appears in an otherwise-valid-looking application ID with
DBusActivatable=yes.  This leads to problems because '-' is indeed valid
in D-Bus names (which suggests that everything should work properly) but
not valid in D-Bus object paths (which suggests that it won't work).

Firm up the range of valid characters permitted in the recommended style
of application IDs (instead of just saying "reverse DNS convention").
Specifically mention that '-' is permitted here.

Specify how to deal with dashes when forming the D-Bus object path for
the org.freedesktop.Application interface.

https://bugs.freedesktop.org/show_bug.cgi?id=95129