mime apps: clarify precedence vs. subtypes
Clarify that we always prefer specific subtypes (like text/html) to the
more generic types (like text/plain) when deciding on associations and
default apps.

This avoids some strange cases where applications that operate on
text/plain are being taken in preference to more-specific applications,
just because they are present in 'higher' directories, or explicitly
listed in defaults files.

https://bugs.freedesktop.org/show_bug.cgi?id=89877