desktop-entry: define an 'iconstring' type
Icons are currently of type 'localestring', which is defined by this
specification as follows:

> Values of type 'localestring' are user displayable, and are encoded in
> UTF-8

Icon names are not user-displayable, and so should not be extracted by
tools like xgettext for translation along with user-displayable strings.
However, we still want tools that load .desktop files to respect the
(rare) cases where icons have been localized. Define a new type,
'iconstring', with the properties we want.

See discussion at https://savannah.gnu.org/bugs/index.php?56543 and the
linked mailing list thread.