thumbnail: Add x-large and xx-large thumbnail spec size
Rather than adding hi-dpi specific text to the thumbnail specification,
we expect code that calls out to thumbnailers to automatically ask for
the right size of thumbnail to be displayed, depending on the level
of integration that the toolkit provides.

The x-large "512x512" size could be used as a 512x512 size for
non-hi-dpi displays, or equivalent to a 256x256 logical size of displays
with a 2x DPI.

The "x*-large" name also has the benefit of being easily extensible
without getting a thesaurus out.

Closes: #15